pyproject.toml
[tool.poetry]
name = "demo2"
version = "0.1.0"
description = ""
authors = ["kotaro-chan <rnaveen28102003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
simple-smtp-mailer = "^0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
